chore: self review fix, rename of the tcp proxy file
