+++ src/activity/exe-unit/index.ts
@@ -1,4 +1,4 @@
 export { ExeUnit, LifecycleFunction, ExeUnitOptions } from "./exe-unit";
 export { Batch } from "./batch";
 export { GolemWorkError, WorkErrorCode } from "./error";
-export { TcpProxy } from "../../network/tcpProxy";
+export { TcpProxy } from "../../network/tcp-proxy";
